test-youtube-video-download/requirements.txt
# Function dependencies, for example:
# package>=version
yt-dlp
google-cloud-storage
hurry.filesize